git-svn-id: https://svtz.ru/svn/graphlabs.site/branches/taskresults@336 e8704808-1a2a-453a-8987-66899a109cbe

+++ GraphLabs.Dal.Ef/GraphLabsDataModel.edmx.sql
@@ -2,7 +2,7 @@
 -- --------------------------------------------------
 -- Entity Designer DDL Script for SQL Server 2005, 2008, 2012 and Azure
 -- --------------------------------------------------
--- Date Created: 06/09/2016 22:31:31
+-- Date Created: 06/09/2016 23:02:20
 -- Generated from EDMX file: C:\Users\Илья\Desktop\taskresults2\GraphLabs.Dal.Ef\GraphLabsDataModel.edmx
 -- --------------------------------------------------
 
@@ -54,7 +54,7 @@ CREATE TABLE [dbo].[Results] (
     [Id] bigint IDENTITY(1,1) NOT NULL,
     [Mode] int  NOT NULL,
     [StartDateTime] datetime  NOT NULL,
-    [Score] int  NULL,
+    [Score] smallint  NULL,
     [Status] tinyint  NOT NULL,
     [Student_Id] bigint  NOT NULL,
     [LabVariant_Id] bigint  NOT NULL
@@ -65,7 +65,7 @@ GO
 CREATE TABLE [dbo].[StudentActions] (
     [Id] bigint IDENTITY(1,1) NOT NULL,
     [Description] nvarchar(100)  NOT NULL,
-    [Penalty] int  NOT NULL,
+    [Penalty] smallint  NOT NULL,
     [Time] datetime  NOT NULL,
     [TaskResult_Id] bigint  NOT NULL
 );
@@ -175,7 +175,7 @@ GO
 CREATE TABLE [dbo].[TaskResults] (
     [Id] bigint IDENTITY(1,1) NOT NULL,
     [Status] tinyint  NOT NULL,
-    [Score] int  NOT NULL,
+    [Score] smallint  NOT NULL,
     [Result_Id] bigint  NOT NULL,
     [TaskVariant_Id] bigint  NOT NULL
 );
