+++ GraphLabs.DomainModel/StudentAction.cs
@@ -24,7 +24,7 @@ namespace GraphLabs.DomainModel
     
         public long Id { get; private set; }
         public string Description { get; set; }
-        public int Penalty { get; set; }
+        public short Penalty { get; set; }
         public System.DateTime Time { get; set; }
     
         public virtual TaskResult TaskResult { get; set; }
